pyproject.toml
[tool.poetry]
name = "viper"
version = "0.1.0"
description = "A package for scanning supply chain vulnerabilities in Open Source LLMs."
authors = ["Rafael Pierre <github.com/weet-ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
cvss = "^3.1"
tqdm = "^4.66.4"
pandas = "^2.2.2"
numpy = "<2.0.0"
packaging = "^24.1"
pyfiglet = "^1.0.2"
colorama = "^0.4.6"
pyspark = "^3.5.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
ruff = "^0.5.0"
ipykernel = "^6.29.5"
nbconvert = "^7.16.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
